Move the VideoReader where it can be reused.
